SaleDaoImpl Test 진행

+++ src/main/java/team/cheese/Controller/SaleController.java
@@ -13,7 +13,6 @@ import team.cheese.domain.SaleDto;
 import team.cheese.service.SaleService;
 
 import javax.servlet.http.HttpSession;
-import java.math.BigInteger;
 import java.util.List;
 
 @Controller
@@ -47,7 +46,7 @@ public class SaleController {
     
     // 게시글 리스트 중 하나를 클릭한 경우
     @RequestMapping("/read")
-    public String read(BigInteger no, Model model, HttpSession session) throws Exception {
+    public String read(Long no, Model model, HttpSession session) throws Exception {
         SaleDto saleDto = saleService.read(no);
 
         System.out.println("sale.jsp로 전달");
