+++ src/main/java/team/cheese/service/SaleServiceImpl.java
@@ -7,7 +7,6 @@ import team.cheese.dao.SaleCategoryDao;
 import team.cheese.dao.SaleDao;
 import team.cheese.domain.SaleDto;
 
-import javax.servlet.http.HttpSession;
 import java.math.BigInteger;
 import java.util.List;
 
@@ -48,7 +47,7 @@ public class SaleServiceImpl implements SaleService {
     }
 
     @Override
-    public SaleDto read(BigInteger no) throws Exception {
+    public SaleDto read(Long no) throws Exception {
         // 판매글 번호를 넘겨 받아서 Dao에서 select로 처리
         SaleDto saleDto = saleDao.select(no);
